+++ src/nova/trame/model/ornl/neutron_data_selector.py
@@ -26,7 +26,7 @@ class NeutronDataSelectorState(DataSelectorState):
     def get_facilities(self) -> List[str]:
         raise NotImplementedError()
 
-    def get_instruments(self) -> List[str]:
+    def get_instruments(self) -> List[Dict[str, str]]:
         raise NotImplementedError()
 
 
@@ -50,7 +50,7 @@ class NeutronDataSelectorModel(DataSelectorModel):
     def get_facilities(self) -> List[str]:
         return natsorted(self.state.get_facilities())
 
-    def get_instruments(self) -> List[str]:
+    def get_instruments(self) -> List[Dict[str, str]]:
         return natsorted(self.state.get_instruments())
 
     def get_experiments(self) -> List[str]:
@@ -59,5 +59,5 @@ class NeutronDataSelectorModel(DataSelectorModel):
     def get_directories(self, base_path: Optional[Path] = None) -> List[Dict[str, Any]]:
         raise NotImplementedError()
 
-    def get_datafiles(self, *args: Any, **kwargs: Any) -> List[str]:
+    def get_datafiles(self, *args: Any, **kwargs: Any) -> List[Dict[str, str]]:
         raise NotImplementedError()